Crate pedersen_scheme

Implementation of Pedersen commitments over an arbitrary curve.

Structs

Commitment

A Commitment is a group element.

CommitmentKey

A commitment key is a pair of group elements that are used as a base to raise the value and randomness, respectively.

Randomness

Randomness used in the commitment. Secret by default.

Value

A secret value. The idea of this datatype is to mark some scalars as secret, so that their use is harder and there is no implicit copy.